Rename a few classes
